depackagify sw plugin, ts
