+++ Report/lighthouse/custom-config.ts
@@ -136,12 +136,12 @@ const defaultConfig = {
     // {id: 'NetworkUserAgent', gatherer: 'network-user-agent'},
     // {id: 'OptimizedImages', gatherer: 'dobetterweb/optimized-images'},
     // {id: 'ResponseCompression', gatherer: 'dobetterweb/response-compression'},
-    {id: 'CustomServiceWorkerGatherer', gatherer: 'custom-lighthouse-service-worker/custom-service-worker-gatherer'},
+    {id: 'CustomServiceWorkerGatherer', gatherer: 'dist/Report/lighthouse/custom-audits/service-worker/custom-service-worker-gatherer'},
     {id: 'WebAppManifest', gatherer: 'web-app-manifest'},
   ],
   audits: [
     'is-on-https',
-    'custom-lighthouse-service-worker/custom-service-worker-audit',
+    'dist/Report/lighthouse/custom-audits/service-worker/custom-service-worker-audit',
     'viewport',
     'metrics/first-contentful-paint',
     'metrics/largest-contentful-paint',
